+++ src/PQ.h
@@ -9,10 +9,10 @@ typedef struct pqueue *PQ;
 PQ      PQinit(int maxN);
 void    PQfree(PQ pq);
 int     PQempty(PQ pq);
-void    PQinsert(PQ pq, int index, int priority);
+void    PQinsert(PQ pq, int index, float priority);
 Item    PQextractMin(PQ pq);
-void    PQchange (PQ pq, int node_index, int priority);
-int     PQsearch(PQ pq, int node_index, int *priority);
+void    PQchange (PQ pq, int node_index, float priority);
+int     PQsearch(PQ pq, int node_index, float *priority);
 void    PQdisplayHeap(PQ pq);
 
 #endif
